requirements.txt
asgiref==3.2.7
cachetools==4.1.1
defusedxml==0.6.0
diff-match-patch==20181111
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.7
django-crispy-forms==1.9.0
django-gmailapi-backend==0.1.0
django-import-export==2.0.2
django-taggit==1.2.0
et-xmlfile==1.0.1
google-api-python-client==1.7.11
google-auth==1.11.2
google-auth-httplib2==0.0.4
httplib2==0.19.0
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.3
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2019.3
PyYAML==5.3.1
rsa==4.0
six==1.15.0
sqlparse==0.3.1
static3==0.7.0
tablib==1.1.0
uritemplate==3.0.1
wfastcgi==3.0.0
xlrd==1.2.0
xlwt==1.3.0
